+++ CHANGELOG.md
@@ -13,6 +13,7 @@
  - handle `edit_url` in `card` and show a button on the right side as in the web version
  - handle `custom_style` for style `div`. Adding the customization from `color_background`, `color_border` and `color_text` fields
  - handle field `load_as_table` in style `entryList`
+ - add `url_cancel` to `formUserInput`
 
 ### Bugfix
  - #2 - fix the validation, `url` is correctly recognized with the capacitor.